requirements.txt
pathlib
datetime
tqdm
pandas
Pillow
piexif